moved new_password to journalist_app/account.py
